updated is.js, refactored jwt code
